Move implementation of HasTypeInfo
